Cosmetic SPECIAL Stats and Stat Check Emotes
Adds SPECIAL stats to the character preferences page. These stats are purely cosmetic and don't affect the gameplay at all. The minimum value of the stats is 1, and the max is 8.  Adds emotes that use these stats to mimic the stat checks present in the games. For example, by using the *strength or *str emote you can do Strength check based on your Strength SPECIAL stat, with fun success and failure messages.